Make reusable skipping `data` checks for management commands
Move overwritten BaseCommand.check() to a separate Mixin class.
